+++ modules/graph_filtering_module/graph_filtering.py
@@ -1,4 +1,5 @@
 import os
+import re
 import sys
 import time
 import concurrent.futures
@@ -329,8 +330,20 @@ class GraphFiltering(QWidget):
         self.logger = logging.getLogger(__name__)
 
     def mask_graph_table_changed(self):
+        self.input_image.setPlaceholderText('')
+        self.input_image.setToolTip('')
         if self.mask_graph_table.rowCount() > 1:
             self.display_results.setChecked(False)
+        elif self.mask_graph_table.rowCount() == 1:
+            mask_path = self.mask_graph_table.get_file_table()[0][0]
+            res = re.match('(.*)'+gf.output_suffixes['segmentation']+'.*$', os.path.basename(mask_path))
+            if res:
+                for ext in gf.imagetypes:
+                    image_path = os.path.join(os.path.dirname(mask_path), res.group(1)) + ext
+                    if os.path.isfile(image_path):
+                        self.input_image.setPlaceholderText(image_path)
+                        self.input_image.setToolTip(image_path)
+                        break
         self.display_results.setEnabled(self.mask_graph_table.rowCount() <= 1)
 
     def get_widgets_state(self):
@@ -411,6 +424,10 @@ class GraphFiltering(QWidget):
 
         if self.input_image.isEnabled():
             image_path = self.input_image.text()
+            if image_path == '':
+                image_path = self.input_image.placeholderText()
+            elif image_path.strip() == '':
+                image_path = ''
         else:
             image_path = ""